backend/requirements.txt
django==5.1
djangorestframework==3.15
django-cors-headers==4.6
psycopg[binary]>=3.2
Pillow>=10.0
